traitement retours certains champs readmodel inutiles pour l'instant
