+++ src/components/shared/Tag/Tag.tsx
@@ -1,0 +1,13 @@
+import Link from 'next/link'
+import { PropsWithChildren, ReactElement } from 'react'
+
+export default function Tag({ children }: PropsWithChildren): ReactElement {
+  return (
+    <Link
+      className="fr-tag"
+      href="/"
+    >
+      {children}
+    </Link>
+  )
+}
