+++ src/use-cases/testHelper.ts
@@ -67,49 +67,7 @@ export function gouvernanceReadModelFactory(
     departement: 'Rhône',
     feuillesDeRoute: [
       {
-        actions: [
-          {
-            beneficiaires: [{ nom: 'Structure 1', roles: ['Porteur'], type: 'Structure' }, { nom: 'Structure 2', roles: ['Porteur'], type: 'Structure' }],
-            besoin: 'EtablirUnDiagnosticTerritorial',
-            budgetGlobal: 50_000,
-            demandesDeCofinancement: [
-              {
-                emetteur: { nom: 'Région Auvergne-Rhône-Alpes', roles: ['Financeur'], type: 'Collectivité' },
-                montantDemande: 15_000,
-              },
-            ],
-            demandesDeSubvention: [
-              {
-                montantAccorde: 35_000,
-                montantDemande: 40_000,
-                type: 'FSE',
-              },
-            ],
-            nom: 'demandeDeSubventionFormation',
-            statut: 'envoyee',
-          },
-          {
-            beneficiaires: [{ nom: 'Structure 1', roles: ['Porteur'], type: 'Structure' }, { nom: 'Structure 2', roles: ['Porteur'], type: 'Structure' }],
-            besoin: 'CoConstruireLaFeuilleDeRoute',
-            budgetGlobal: 50_000,
-            demandesDeCofinancement: [
-              {
-                emetteur: { nom: 'Région Auvergne-Rhône-Alpes', roles: ['Financeur'], type: 'Collectivité' },
-                montantDemande: 15_000,
-              },
-            ],
-            demandesDeSubvention: [
-              {
-                montantAccorde: 35_000,
-                montantDemande: 40_000,
-                type: 'FSE',
-              },
-            ],
-            nom: 'demandeDeSubventionFormation',
-            statut: 'validee',
-          },
-        ],
-        beneficiaires: [
+        beneficiairesSubvention: [
           {
             nom: 'Structure 1',
             roles: ['Co-porteur'],
@@ -137,7 +95,6 @@ export function gouvernanceReadModelFactory(
         montantSubventionAccorde: 5_000,
         montantSubventionDemande: 40_000,
         montantSubventionFormationAccorde: 5_000,
-        montantSubventionFormationDemande: 40_000,
         nom: 'Feuille de route inclusion 1',
         porteur: { nom: 'Préfecture du Rhône', roles: ['Co-orteur'], type: 'Administration' },
         totalActions: 3,
